+++ json/src/main/java/org/jdbi/v3/json/internal/JsonArgumentFactory.java
@@ -23,6 +23,7 @@ import org.jdbi.v3.core.config.ConfigRegistry;
 import org.jdbi.v3.core.internal.JdbiOptionals;
 import org.jdbi.v3.core.qualifier.QualifiedType;
 import org.jdbi.v3.core.statement.UnableToCreateStatementException;
+import org.jdbi.v3.json.EncodedJson;
 import org.jdbi.v3.json.Json;
 import org.jdbi.v3.json.JsonConfig;
 
@@ -33,20 +34,17 @@ import org.jdbi.v3.json.JsonConfig;
 public class JsonArgumentFactory implements ArgumentFactory {
     private static final String JSON_NOT_STORABLE = String.format(
         "No argument factory found for `@%s String` or 'String'",
-        Json.class.getSimpleName()
+        EncodedJson.class.getSimpleName()
     );
 
     @Override
     public Optional<Argument> build(Type type, Object value, ConfigRegistry config) {
-        if (String.class.equals(type)) {
-            return Optional.empty();
-        }
         String nullableJson = value == null ? null : config.get(JsonConfig.class).getJsonMapper().toJson(type, value, config);
         String json = "null".equals(nullableJson) ? null : nullableJson; // json null -> sql null
         Arguments a = config.get(Arguments.class);
         // look for specialized json support first, revert to simple String binding if absent
         return Optional.of(JdbiOptionals.findFirstPresent(
-                () -> a.findFor(QualifiedType.of(String.class).with(Json.class), json),
+                () -> a.findFor(QualifiedType.of(String.class).with(EncodedJson.class), json),
                 () -> a.findFor(String.class, json))
                 .orElseThrow(() -> new UnableToCreateStatementException(JSON_NOT_STORABLE)));
     }